Refactor backend structure file
