--- backend/.env.template
@@ -1,2 +1,0 @@
-DATABASE_URL=mysql+pymysql://root@localhost:3306/ your_database_name
-SECRET_KEY=put_secret_key_here
