+++ backend/app/.env.template
@@ -1,0 +1,2 @@
+DATABASE_URL=mysql+pymysql://root@localhost:3306/your_database_name
+SECRET_KEY=put_secret_key_here
