+++ backend/app/config.py
@@ -1,0 +1,9 @@
+from pydantic import BaseSettings
+
+class Settings(BaseSettings):
+    DATABASE_URL: str
+
+    class Config:
+        env_file = ".env"  
+
+settings = Settings()
