+++ backend/app/models/user.py
@@ -1,0 +1,38 @@
+from datetime import datetime, timezone
+from sqlalchemy import Integer, String, DateTime
+from sqlalchemy.orm import Mapped, mapped_column
+
+from app.db.base import Base
+
+
+class User(Base):
+    __tablename__ = "users"
+
+    id: Mapped[int] = mapped_column(
+        Integer, primary_key=True, index=True
+    )
+    role: Mapped[int] = mapped_column(
+        Integer, nullable=False
+    )
+    first_name: Mapped[str] = mapped_column(
+        String(15), nullable=False, index=True
+    )
+    last_name: Mapped[str] = mapped_column(
+        String(15), nullable=False, index=True
+    )
+    user_email: Mapped[str] = mapped_column(
+        String(40), nullable=False, unique=True, index=True
+    )
+    user_password: Mapped[str] = mapped_column(
+        String(80), nullable=False
+    )
+    registered_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        default=lambda: datetime.now(timezone.utc),
+        nullable=False,
+    )
+    edited_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        onupdate=lambda: datetime.now(timezone.utc),
+        nullable=True,
+    )
